Use bufio and independent buffers for memd body reading.
This actually improves performance in most cases (but creates
some additional GC pressure, though this could be resolved in
a future commit by using a pool of value buffers).  This is
also a prerequisite change for offloading JSON decoding to the
caller goroutines rather than doing it on the IO goroutines.

Change-Id: I59ae39b824255797cabda34eb1db32b6c88d16a7