Add new flags for set_meta/del_meta operations
The current SkipConflictResolution flag has been renamed to
ForceWithMetaOp. The behavior has not changed on the server side,
but turns out that that flag allowed writes to active, replica, and
pending vbuckets. A new flag has been added that skips conflict
resolution, but only allows writes to the active vbucket. This
change should be ok since thse flags are internal only.

Change-Id: I62fe70f0a2689912cd3706778d6810cc25b7ddc4
Reviewed-on: http://review.couchbase.org/80480
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>