GOCBC-213: Replaced ErrInternalError use with ErrCliInternalError.
Motivation
----------
In various places throughout the Go SDK, the kv-oriented ErrInternalError
has been used to signal internal errors.  This causes confusion when
debugging as it is not possible to immediately understand where the source
of any particular internal error comes from.

Changes
-------
Added a new ErrCliInternalError and updated all uses of ErrInternalError
within the Go SDK to utilize the new error constant.

Change-Id: I534891bbdd35b41a0d7b843b7e6b35270ff548d1
Reviewed-on: http://review.couchbase.org/80615
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Brett Lawson <brett19@gmail.com>