GOCBC-215: Refactor errors objects to support extra information.
Motivation
----------
The current kv error objects simply store the error code of the
error that occured.  This does not allow for dynamic error messages
or additional contextual information that may be available about
the error.  Since dynamic error information will be implemented
soon, the errors need to be refactored to allow additional data
to be stored.

Changes
-------
Refactored memdError into a public KvError structure containing
the code, message and name of any error that has occured.  We
additionally have exposed the status codes for memcached to allow
developers to test the type of error that occured in the future
when dynamic error creation is implemented.

Change-Id: Idcf75be60bf5602111396261d1a4ef12789b90c5
Reviewed-on: http://review.couchbase.org/80617
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>