GOCBC-220: Fixed tests sporadically (racy) failing.
Changes
-------
The Stats test was not initiating a signaller wait, this meant
that the callback for the stats would block waiting for someone
to receive the completion event on the channel.  This would block
the callback loop for the last responding server for the Stats
operation.  Any subsequent operations to this server would never
receive results, as the servers loop would be blocked.

Change-Id: I26772ab757f14153ed208d11f2f2ef16e530d1d7
Reviewed-on: http://review.couchbase.org/80756
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>