GOCBC-235: Always SELECT_BUCKET on Spock clusters.
Motivation
----------
Due to the distributed nature of Couchbase Server, the
independance of authentication and bucket selection in
Server 5.0 causes a race during node startup which can
lead to an unrecoverable state.

Changes
-------
We always perform SELECT_BUCKET when connecting to any
serve which supports it.  This ensures the SDK is notified
via a NOT_FOUND (transient) or NO_ACCESS (permanent) error.

Change-Id: I51f783a14f42111e07df4a13c0aa80bd23fd4504
Reviewed-on: http://review.couchbase.org/83868
Reviewed-by: Adam Fraser <adam.fraser@couchbase.com>
Tested-by: Adam Fraser <adam.fraser@couchbase.com>