GOCBC-238: Ensure that DCP buffer is flushed before flushing op list.
Motivation
----------
Due to how DCP handles packets, and internally handles closing
the request once it has received an end packet, a race condition
exists between client shut-down and receiving a stream-end from
the server.  This manifests as receiving two stream-end messages
if the server is closed as a stream-end packet is received.  This
occurs because the stream-end message can be sitting in the dcp
queue waiting to be processed when the connection closes.  The
connection closing causes an ErrNetwork to be sent to the op list
and then later the buffer is processed, sending an additional
stream-end message.

Changes
-------
Added synchronization to ensure the dcp buffer is flushed prior to
attempting to flush the op list and completing the client cleanup.

Change-Id: Ie033c19e5f708f85890de90435617af263889b59
Reviewed-on: http://review.couchbase.org/84389
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>