GOCBC-244: Fix bug causing IPv6 addresses to become corrupted.
Motivation
----------
In testing for IPv6 support, we determined that the net.SplitHostPort
function of Go will strip away the backets which indicate an IPv6
address within our host-port pairs.

Changes
-------
Implement our own version of host/port splitting which piggybacks
on the existing net.SplitHostPort, but restores the brackets which
were removed.

Change-Id: If20051eac8efd4c0f08e606d0cd62a6b4948f5da
Reviewed-on: http://review.couchbase.org/85351
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>