GOCBC-249: Rewrote stats function to fix various issues.
Motivation
----------
The stats call in gocbcore has consistently been a pain point in
our implementation due to the fact that it frequently would
return inconsistent data without warning and had various bugs due
to it being the only multi-server aggregation method in the SDK.
Recently some inconsistency in how the operations get completed in
the case of partial errors was discovered prompting this issue.

Changes
-------
Rewrote the stats call to take advantage of some new functionality
in our routing layer and cleaned up many of the race conditions,
panics and inconsistencies that existed in the function.

Change-Id: I6ee840acb51283d51758df9192036634715bbe52
Reviewed-on: http://review.couchbase.org/85652
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>