GOCBC-251: Implement KeyToServer for Memcached Buckets (Ketama)
Motivation
----------
While debugging some inter-SDK compatibility issues, we discovered
that performing KeyToServer when using a Memcached bucket does not
work at all.  Ketama is still able to resolve a key to a server,
so it is expected to work.

Changes
-------
Implemented missing Ketama lookup to add support for Memcached
buckets.

Change-Id: Ib7de19b270f8ffdf73c8c5cba6033658e5cc2f76
Reviewed-on: http://review.couchbase.org/85788
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>