GOCBC-252: Fix race detected in memdPipeline code.
Motivation
----------
Two methods of the memdPipeline took non-pointer receivers.
This caused the memdPipeline to be copied to the stack by
invokers, which caused some of the mutable fields of the
pipeline to be read and a race condition detected (though
it wasn't a real race).  A future change will correct some
loss of immutability in the code.

Changes
-------
Modified the two methods to be pointer receivers.

Change-Id: Ia1fd4e08c4d4fea64def0bb72d28109f4b76f946
Reviewed-on: http://review.couchbase.org/85860
Reviewed-by: Adam Fraser <adam.fraser@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>