GOCBC-245: Added new health-check related APIs.
Motivation
----------
As part of the new Health Check APIs being implemented in the SDK,
We are implementing the ability to ping all services and generate
a report regarding the connectivity of the SDK to the cluster to
aid in diagnostics and monitoring.  Additionally, we are providing
APIs to fetch the state of all active connections.

Implemented a new Ping method on the Agent structure which dispatches
a NOOP to all nodes and aggregates the results. Additionally
implemented new locking to the connection mangement code, and built
a function which can iterate the active connection muxer to build a
list of active connections with diagnostics information.

Change-Id: I1ed285c4fb14286d3078c88325242bb0cacc7326
Reviewed-on: http://review.couchbase.org/85959
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>