GOCBC-256: Expose authentication provider hooks.
Motivation
----------
Due to the ongoing use of the Go SDK by teams internally
within Couchbase.  There is a need for enabling developers
to utilize the internal server authentication method.

Changes
-------
Implemented an AuthProvider interface which provides a
callback for fetching specific credentials based on
endpoint, service, etc...  Updated all locations within
the codebase which use credentials to use the provider.

Change-Id: Ifd36a45c4e77e89d694be7deaaa3df9b50acf4fe
Reviewed-on: http://review.couchbase.org/87072
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Gautham B A <gautham.bangalore@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>