GOCBC-267: Implemented support for network compression.
Motivation
----------
In order to reduce the amount of bandwidth consumed, Couchbase
Server now supports compression of document bodies when
sending and receiving them over the network.

Changes
-------
Implemented hooks in the core memcached client to enable the
compression datatype to be understood and the compressed
document contents automatically decompressed before
processing.

Change-Id: I6e078215e45202a0550b54115067549cfd862f09
Reviewed-on: http://review.couchbase.org/90167
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>