GOCBC-292: Check for malformed packets before reading.
Motivation
----------
There have been a few instances of the SDK encountering a
packet which appears to be malformed, causing it to crash.
Ideally, the SDK should be verifying the contents of the
packets it receives prior to its attempts to read them to
enable graceful handling of these issues instead.

Changes
-------
Added additional validation on packets that were not
previously validated prior to reading.

Change-Id: Iac797d5ae937b34872ca0589e53006a0da0af0ae
Reviewed-on: http://review.couchbase.org/93080
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Jaekwon Park <jaekwon.park@couchbase.com>