GOCBC-159: Added connstr options for http client tuneables.
Motivation
----------
Under some workloads, a large number of connections are established
for HTTP requests (such as for N1QL).  Due to the TIME_WAIT
implementation of linux socket handling, this can cause the pool
of local ports to become exhausted and new connections to fail.

Changes
-------
Added tuneable properties to control the pooling tuneables of the
HTTP transport used by the SDK to allow developers to tune the
number of connections kept in the pool, reducing the number of
discarded connections.

Change-Id: Ice0386363bb265067fae827c491ca5bc91c29549
Reviewed-on: http://review.couchbase.org/93782
Tested-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Adam Fraser <adam.fraser@couchbase.com>