GOCBC-305: Added zombie response aggregation and logging.
Motivation
----------
Implementation of additional traceability in the SDK involves
the ability to aggregate and log information about responses
which are received from the server after the operation it was
attached to has already timed out.

Changes
-------
Added a zombie log to the agent and implemented ordered
insertion and periodic logging of the information which was
aggregated.

Change-Id: I37086c012f1d11cc5bf31177d14107cf8fb66276
Reviewed-on: http://review.couchbase.org/94437
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>