GOCBC-303: Expose settings for compression min_size/min_ratio
Motivation
----------
We should allow users to configure the minimum document size
for a document to be compressed before sending to the server.
We should also allow the user to configure the minimum ratio of
compressed to original document size for the compressed document
to be sent.

Changes
-------
Expose compression_min_size and compression_min_ratio options
via the connection string. Check against these options
before a) compressing the document and b) sending it.

Change-Id: I8a1db44bc0715b6b0c7db3a74158ea95be1c066b
Reviewed-on: http://review.couchbase.org/94563
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>