GOCBC-316 Fix panic due to nil pointer on span Finish
Motivation
----------
It's possible for a request to timeout and the corresponding
response to return at roughly the same time. When this happens it's
possible for calling Finish on cmd and net spans to race and due to
stopCmdTrace/stopNetTrace setting the spans to nil this means
that the timeout context can nil pointer. This can also occur
when the timeout value is so small that the request isn't
even sent.

Changes
-------
Added a Mutex to memdqpackets and lock the request on processing
a response or a timeout..

Change-Id: I6e068c83c2db13e2818ffaa11628d586006ff39a
Reviewed-on: http://review.couchbase.org/95633
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>