GOCBC-319 Prevent stopCmdTrace from being called twice on one request
Motivation
----------
Under circumstances when a request contains an error that is either
a NMV or kv error map error then stopCmdTrace can be called twice
for the same span. The effect of this is that the logs can get filled
with spam.

Changes
-------
Alter where stopCmdTrace is called so that it is not called twice
for the same trace.

Change-Id: I7fff5c37301c0ff4e81a38e464c06f71ae553059
Reviewed-on: http://review.couchbase.org/95973
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>