GOCBC-327: Fix data race on zombie logger
Motivation
----------
A data race can occur between the zombie logger loop and
logging a new zombie response. This is due to the zombieOps
slice being modified by logging a new zombie response at the same
time as the logger loop tries to access the capacity of the slice.

Changes
-------
Alter the zombie logger function to take a sampleSize corresponding
to the zombieOps capacity  as a parameter.

Change-Id: I56c0f941264c7421714c11da0d9947f11b90a15c
Reviewed-on: http://review.couchbase.org/96664
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>