GOCBC-328: Fix a data race on draining a client.
Motivation
----------
A data race is possible on a memdClient's memdOpList when
the client is closed from the remote side.

Changes
-------
Change the memdQRequest waitingIn owner to be memdClient rather
than the memdOpMap. Remove locks from the opMap and move them up
a level to the memdClient.  Additionally added checks to ensure
the client is not closed before mutating the opMap.

Change-Id: I4fa945e2ea32fdf03179242986ca65d4978c634d
Reviewed-on: http://review.couchbase.org/96675
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>