GOCBC-334: Fail testing for leaked goroutines.
Motivation
----------
There have been cases where minor bugs in the SDK
have caused the number of goroutines to steadily
increase, but this is frequently lost among the
noise of running the tests.

Changes
-------
Added code to keep track of the number of running
goroutines before and after the tests run, along
with code to fail the testing if any goroutines
are left running.

Change-Id: I7bf8efec5aef7b977e26a6f8b1f2a1530d5e30a8
Reviewed-on: http://review.couchbase.org/96859
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>