GOCBC-332: Close testing agents when completed.
Motivation
----------
In order to improve testing behaviour, it is intended
that checking to ensure that we do not leak goroutines
is added.  Currently many goroutines are leaked by our
tests as the tests do not actually shut down the
agents that are used gracefully.

Changes
-------
Added post-test handling to shut down all Agents.

Change-Id: I17873f83040b0495ab966f1cee5f7099f2a268e2
Reviewed-on: http://review.couchbase.org/96858
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>