GOCBC-345: Make sure kv service is only enabled if in nodes list.
Motivation
----------
The SDK needs to iterate the nodesExt as part of the config because
some global services (i.e. N1QL) are not to be found in the
traditional "nodes" list of the config. With KV, this turns out
to be a problem during rebalance-in scenarios.

Especially during delta-node recovery and memcached buckets, the
problem amplifies because nodesExt holds all nodes that are part
of the cluster with their services, while the nodes list only
holds the nodes that are currently active for the given bucket.

If the SDK only uses nodesExt (which it does now), it tries to
connect to kv buckets immediately as they become available, even
if they are not active for this bucket yet. For couchbase and
ephemeral buckets this is not a big issue because there the
movement of vbuckets gradually moves over the load, but with
memcache buckets as soon as the node pops up in the config
it is part of the ketama ring and ops are dispatched to it
right away.

Changes
-------
The part of the config loading where we iterate nodesExt now
explicitly checks if a corresponding entry in the nodes list is
available. If not then the node is not added to the kvServerList,
it can still be added to the list for other services.

Change-Id: Iebfc5d44838c2ea6e89ef4d2a4d77046ec3c7fa6
Reviewed-on: http://review.couchbase.org/99034
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Charles Dixon <chvckd@gmail.com>