GOCBC-352: Don't log invalid config on initial cluster connect
Motivation
----------
When connecting to a cluster under some circumstances then an
invalid config message is logged. This isn't an issue on initial
cluster connect but is when later updating the config.

Changes
-------
Check if the config build is being performed on an initial cluster
connect and, if so, then don't log the message.

Change-Id: Ia3407c56c131045fefea48c25e059d5a1542c264
Reviewed-on: http://review.couchbase.org/100278
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>