GOCBC-363: Prevent Stats call from hanging
Motivation
----------
When using the Stats call it's possible to enter a scenario where
the call can hang. Given 3 servers so 3 requests from gocbcore,
if 2 responses come back in good time but 1 does not then a timeout
will occur. During this timeout gocb cannot complete the call as it
does not know that the operation has multiple requests and as
cancelling the operation returns false it appears that the corresponding
response has just arrived and so waits for it to complete. What is
actually happening is that 2 responses have arrived and 1 never will
so waiting for the operation to complete hangs.

Changes
-------
Add a completed count to multiOp and on Cancel ensure that completed
== failed to cancel (because they've already completed).

Change-Id: I3f6e94cc19b61b0db06347e13eccc350996b1bf2
Reviewed-on: http://review.couchbase.org/101223
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>