GOCBC-371: Prevent tests from hanging when a test times out
Motivation
-----------
If a test times out when running the tests against a real server then when the tests complete the closing of the globalAgent hangs indefinitely.

Changes
-------
Added a check for when tests are complete and when complete do not
call any request callbacks. The tests have already failed by this point.

Change-Id: I222950f0af4facf3494acd130529e7037487d9d2
Reviewed-on: http://review.couchbase.org/101643
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>