GOCBC-380: Fix initial route config to not ignore networkType
Motivation
----------
The SDK currently ignores the value specified by the user and instead uses the heuristic in all cases.  This is an error and the heuristic should only be used if there is no value specified, or the value is 'auto'.

Changes
-------
Updated agent.buildFirstRouteConfig to check if the user has set
a network type and if so to use the type that they have set.

Change-Id: I04cb138740925ac8037c85cf6615fbf73c2808fd
Reviewed-on: http://review.couchbase.org/102423
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>