GOCBC-381: Split BasicOps and GetReplica Tests
Motivation
----------
Test are intermittently failing on the CI system and it looks like
the reason is the key set in BasicOps is not yet replicated over to
the node requested in GetReplica in the same test.

Changes
-------
Split the GetReplicas out of the BasicOps test and add an observe
in a retry loop after doing the Set, before doing GetReplica.

Change-Id: I721ed4a1d3c1a005d201693a5edb447058e0a390
Reviewed-on: http://review.couchbase.org/102590
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>