GOCBC-385: Add support for enable_expiry_opcode dcp control
Motivation
----------
Couchbase server now supports differentiating between deleted and
expired documents for DCP. Whilst gocbcore already supports this
there is a new DCP control opcode that must be negotiated for it
to be enabled.

Changes
-------
Add the ability to enable sending this new control during agent
initialization. Change the signature of initFn to accept an agent
pointer so that a flag can be set on the agent to signal whether
or not the negotiation succeeded. By not erroring this means that
clients can continue to work with older server versions but provide
a way to check if expiry is supported.

Change-Id: Ib660d02813a1da1b6313f85ced3133fead626fb3
Reviewed-on: http://review.couchbase.org/104248
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>