GOCBC-397: Automatically retry KV ops if collection unknown
Motivation
----------
When an operation is performed against a known collection
(so has a valid cid), if it fails with collection unknown
(e.g. because the collection has been deleted and recreated)
then we should automatically try to update the cid and try again.

Changes
-------
Replace logic that performs a get cid before an op if
collection is unknown. Instead, if a collection is
known to be unknown then perform a get cid and add
any ops for that collection to a queue which is drained on
get cid response.

Change-Id: I88eff8916c75deaed4e9e870d5f8e600b26e7936
Reviewed-on: http://review.couchbase.org/105830
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>