GOCBC-413 - Fix race causing pipeline client to hang indefinitely
Motivation
----------
If a request times out (from gocb or syncclient) whilst it is being
processed for sending then it can result in SendRequest returning an
error and the pipelineclient attempting to shut down. In this scenario it is
possible for the signal in the timeout handler to have already been
processed so the call to req.Callback blocks whilst the signal write
waits for a read.

Changes
-------
Change the pipeline client SendRequest error handler to call
req.tryCallback rather than req.Callback. If the request has already
been cancelled (by a timeout) then this prevents the callback from
being called.

Change-Id: Ie0126a251eec72045f5641c4a139db0e6300f1ca
Reviewed-on: http://review.couchbase.org/106940
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>