GOCBC-411: Create GetAnyReplicaEx and move GetReplicaEx to GetOneReplicaEx
Motivation
----------
Gocb requires a GetAnyReplica function that can also query the active server
as well as the ability to get a replica from a specific server.

Changes
-------
Split the any logic out from GetReplicaEx and put it into
GetAnyReplicaEx, which also calls GetEx to query the active node.
Create GetOneReplicaEx for querying a specific node.

Change-Id: I9bdcba4ca4d138321cf51b33bfb26577fa002055
Reviewed-on: http://review.couchbase.org/107709
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>