GOCBC-431: Fix issue where alternate addresses ports not set
Motivation
----------
Within the alternateAddresses section of the cluster config map
it's possible for the ports to be missing in the JSON. This happens
in the case of a publically addressable Kubernetes cluster (using
public DNS).

Changes
-------
Make the Ports field of cfgNodeAltAddress an omittable pointer,
check if it is set before assigning it as the ports to use.

Change-Id: Iaec94349e2005928569356fc4235afbd9f911703
Reviewed-on: http://review.couchbase.org/107871
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>