GOCBC-505: Add support for GCCCP
Motivation
----------
As of 6.5 the server supports using connections that haven't had select
bucket run on them. This means that things like queries can be run
without needing to select bucket first.

Changes
-------
Add a cluster level config struct and a gcccp poller. Rewrite
bootstrap to include all of the scenarios that can
occur including no bucket being set on the agent. Move where auth
is performed to also be a part of bootstrap.

Change-Id: I11004876afae79e38a0fb1a4d515932fa92c556c
Reviewed-on: http://review.couchbase.org/112018
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>