Update racey tests and fix a couple of test issues
Motivation
----------
Occassionally alternate addresses tests can race against the
agent performing routing updates. The collection retry test
attempts to run even if no or default collection are set but
server version is 6.5

Changes
-------
Update the alternate addresses tests to run against an agent that
isn't connected to a cluster. Don't run the collection retry test
unless an actual collection is specified. As a bonus also  add the
ability to disable the logger which can be useful in tracking down
failures in repeated tests locally.

Change-Id: I807431bf62113ca73921e733af511e7f67087a1b
Reviewed-on: http://review.couchbase.org/113615
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>