GOCBC-589: Prevent nil pointer on dcp stream error
Motivation
----------
The OpenStream method can optionally accept a StreamFilter, upon
receiving an error the handler function assumes that the stream
filter is not nil which can cause a nil pointer deference.

Changes
-------
Upon receiving an error check if the filter is nil, if so then
set the stream id to the no stream filter id const rather than
attempting to get it from the filter.

Change-Id: Ia412700ffd8b2d3c12d69cf07497ea74f72950b7
Reviewed-on: http://review.couchbase.org/114534
Reviewed-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>