GOCBC-593: Fix dcp test races
Motivation
----------
Several points in the dcp tests are racey and triggering the
race detector.

Changes
-------
Update the observer to use atomics to increment counts and also
use atomics to access those counts from outside the observer.
Use a lock for adding stream ops to the ops list as we add
them concurrently.

Change-Id: I10487f85fd08c8398d689105c1d26452589251f5
Reviewed-on: http://review.couchbase.org/114536
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>