GOCBC-594: Update stats call to handle specific vbuckets
Motivation
----------
The DCP stat call 'dcp-vbtakeover' is a unique stat call
because it accepts arguments. The first argument being a
vBucket ID, therfore, this request *must* be sent to the
server which has the requested vBucket.

Changes
-------
Added an empty StatsTarget interface and option for the StatsEx call.
Also created a VbucketIdTarget struct that can be used for targetting
a specific vbucket.

Change-Id: Ic2623226cd1ae3584bb51dec88427c7983254b57
Reviewed-on: http://review.couchbase.org/114870
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>