GOCBC-591: Add redirect handling to http
Motivation
----------
When the server responds to a http request with a 302 then Go will
resend the request to the server within the 302. As per http spec
it does this without copying any headers from the previous request.
This means that auth is missing from the redirected request and so
it fails. We need to ensure that auth is present on that request.

Changes
-------
Add an implemention for CheckRedirect to the http client. If present
then this will copy over the Authorization header.

Change-Id: I4904c8a74cfbec4bd3a39d05b669e0e96a83c95f
Reviewed-on: http://review.couchbase.org/115040
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>