GOCBC-591: Add redirect handling to http
Motivation
----------
When the server responds to a http request with a 302 then Go will
resend the request to the server within the 302. As per http spec
it does this without copying any headers from the previous request.
This means that auth is missing from the redirected request and so
it fails. We need to ensure that auth is present on that request.

Changes
-------
Add an implemention for CheckRedirect to the http client. If present
then this will copy over the Authorization header.

Change-Id: I97193eb5df6da7993ebdaf2656f0d920de8572ca
Reviewed-on: http://review.couchbase.org/115045
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>