GOCBC-615: Add CollectionId to options blocks
Motivation
----------
Some users (like the tools team) need to be able to set the
Collection ID for an operation directly rather than allowing
us to manage IDs.

Changes
-------
Add CollectionId option to all option blocks that support collections.
If the collection ID is set then just send the request with the
provided collection ID instead of consulting the cache.

Change-Id: Ifbaae81d9d1184fa38b7e122cb8b3d073ded038d
Reviewed-on: http://review.couchbase.org/115261
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>