GOCBC-614: Alter auth func signatures
Motivation
----------
We expose auth function as returning channels which is a bit of an unusual
interface. Instead, we should them to take callbacks as parameters.

Changes
-------
Update AuthFunc to take 2 callbacks rather than return 2 channels.
Add a new authFunc which has the old signature. If AuthFunc is set
then wrap it with a function that returns an authFunc during agent
creation. Update all of the public Sasl functions so that they also
change signature to match. Update the getNextAuth function to convert
the callbacks to channels.

Change-Id: I9fc6421a0b26c5ef3bd0cc9be9b1e86571831c69
Reviewed-on: http://review.couchbase.org/115317
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>