GOCBC-656: Add support for tracing
Motivation
----------
We need to add support for ThresholdLogging which means that
gocbcore once against needs to add support for tracing. This
support should be added as volatile as it WILL be changed when
the opentelemetry API stabilises.

Changes
-------
Added volatile support for tracing across the SDK by exposing
a RequestTracer interface. Default to a noop tracer.

This reverts commit 1be848e673c23c48e41b17ca112de8cda1b8f553.

Change-Id: I4c1eeb19b5862efe4927cb78594781fa1149fbc2
Reviewed-on: http://review.couchbase.org/116729
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>