GOCBC-663: Prevent panic in the zombie logger
Motivation
----------
A rarely occuring panic happens within the zombie logger when
the capacity of the zombie ops slice is 0. The only place that
this can occur is on startup.

Changes
-------
Move enabling the zombie logger on the agent until after connecting.
Add a check to the zombie logger record function to ensure that
the capacity is greater than 0.

Change-Id: I57c28222ff648d869e84d2ffca4f56212c207123
Reviewed-on: http://review.couchbase.org/117109
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>