GOCBC-671: Fixed Stats calls when nodes error in response.
Motivation
----------
The implementation of the Stats response handling required the
response object in order to correctly assign the result of the
operation to the return value.  In some cases the response will
be nil, causing a panic.

Changes
-------
Adjusted the handling such that the response is no longer
needed as part of the core handling.

Change-Id: Id9a7c74c9ea8da1c4196b71ce999546aabb743bd
Reviewed-on: http://review.couchbase.org/118101
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>