GOCBC-682: Correctly setup auth if initial connection is over http
Motivation
----------
If we initially setup the connection to a cluster (for cluster
config fetching, via http looper) then we do not correctly setup
auth for all following memcached connections. This causes select
bucket on those connections to always fail.

Changes
-------
Correctly setup the authHandler as a part of the first function
callback within http looper.

Change-Id: I8dcc94aaaa859759284ef582994364e595a9042c
Reviewed-on: http://review.couchbase.org/121716
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>