GOCBC-774: Fix DCP buffer acknowledgement amount
Motivation
----------
With collections enabled DCP streaming packets now contain
extra information stored in the framing extras etc... We do
not take this into account when we calculating the size to
send as a part of bucker acknowledgement requests. This can result
in DCP streams hanging whilst the server waits for the client to
be ready.

Changes
-------
Return the size of the packet at ReadPacket time and bubble that
through to the ack sending so that we always know the size for
sure.

Change-Id: I2019c561653223c463bb3b539fe548eac0c1f410
Reviewed-on: http://review.couchbase.org/121722
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>