GOCBC-759: Prevent context cancel from closing active connections
Motivtion
---------
We use context package to set deadlines on http requests. As a part
of this we are using defer cancel() which sometimes (such as a
streaming response to a query) leads til closing active connections
early.

Changes
-------
Only call cancel if we return in an error state from the exec http
function. This is safe as the parent context has no cancel function
associated and so no extra go routines are spun up to leak.

Change-Id: I2f5a9db3deee152c2a40717c2968efb0df1dd1ef
Reviewed-on: http://review.couchbase.org/121747
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>