GOCBC-779: Fixed Stats calls when nodes error in response.
Motivation
----------
The implementation of the Stats response handling required the
response object in order to correctly assign the result of the
operation to the return value.  In some cases the response will
be nil, causing a panic.

Changes
-------
Adjusted the handling such that the response is no longer
needed as part of the core handling.

Change-Id: I8c203894bf61d79d773a3adc3493e50e0d89e927
Reviewed-on: http://review.couchbase.org/121938
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>