GOCBC-780: Prevent errmap tests from failing
Motivation
----------
We setup the mock to respond to requests with error map info.
Occasionally a cccp request gets in whilst we're running the test
which causes the test to fail.

Changes
-------
Add a channel to the agent which can be used to set the paused
state of the cccp looper.

Change-Id: I9133e7c0048d94764fe05b5fcb485907e7213de3
Reviewed-on: http://review.couchbase.org/121941
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>