GOCBC-765: Try to parse connection string timeouts as durations
Motivation
----------
Connection string options which are times are currently parsed as
milliseconds. It would be nice if these were parsed as durations
(which default to milliseconds if unspecified).

Changes
-------
Updated connection string parameters which are timeouts so that
they can be specified as durations falling back to milliseconds.
Also add tests for agent config.

Change-Id: If6eb25b93eb3c630907ef50c35ae1a5141e7e1a0
Reviewed-on: http://review.couchbase.org/121983
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>