GOCBC-682: Ensure that connections use the correct credentials
Motivation
----------
When we initially setup authentication we tie the credentials that
we use for all further auth to those of the first endpoint tried.
This fails when different endpoints have different credentials.

Changes
-------
Slightly refactor how we perform authentication so that each time
we bootstrap on a connection we fetch the credentials for that
individual endpoint.

Change-Id: I91fe8897f8fc52a44ac8e7adeb05bc300674de43
Reviewed-on: http://review.couchbase.org/122030
Reviewed-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>