GOCBC-662: Reorder subdoc opts if xattrs are present
Motivation
----------
The server requires that any subdoc xattr ops are at the beginning
of the ops list. If the user provides this list out of order we should
reorder it for them and then reorder it back when we get the result.
This make it so that ContentAt works as expected.

Changes
-------
Added logic to reorder subdoc ops as necessary and then reorder them
back to how they were when we build the result.

Change-Id: I3bb7a50f21311b093ef91592eb6a6857fb06c5d5
Reviewed-on: http://review.couchbase.org/122704
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>