GOCBC-812: Add missing properties to KeyValueError
Motivation
----------
The error handling RFC states that last dispatched to/from and
last connection ID should be included in errors. We have all of the
scaffolding there, we just don't put the information into the errors.
We should do that.

Changes
-------
Add last dispatched to/from and last connection ID to KeyValueError
and populate it along with other error properties.

Change-Id: I24c22d97e039e1e5289abb0bca03717850de3b74
Reviewed-on: http://review.couchbase.org/122838
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>