Split config management into components
Motivation
----------
We should simplify the distinction between gcccp and cccp loopers
by combining them to reduce code duplication. We should also combine
cfgBucket and cfgCluster into one as cfgBucket is a superset of
cfgCluster anyway and there's loads of code duplication there too.
This will then mean that we can make configuration components
for handling ns config and route configs. We will also be able
to make components to own kv and http "muxers".

Changes
-------
Remove the gcccp looper and add a stop signal to the cccp looper.
Remove the cfgCluster and update BuildRouteConfig to support
a config that does not have a bucket.
Add components for handling ns config and route config.
Add http and kv muxers for holding routing state.

Change-Id: I7c0930f230bc3d66a14092fb76e58c6c3832e40f
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125366
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>