GOCBC-819: Return connect certificate errors to the user
Motivation
----------
When we connect to a node we immediately fail and return the error
if the error is an access or auth error. We do not return the error
if the error has originated from a x509 issue. These are technically
access/auth issues so we should we return them.

Changes
-------
On dial fail add additional checks for whether errors are x509 errors,
if they are then return them to the user.

Change-Id: Iad0c8d20ef609e35b754aac9fdb529781e24455c
Reviewed-on: http://review.couchbase.org/123131
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Sarath Kumar Sivan <sarathkumarsivan@couchbase.com>
Reviewed-by: Sarath Kumar Sivan <sarathkumarsivan@couchbase.com>