GOCBC-823: Remove onCompletion callback from memdqrequest
Motivation
----------
When a request is cancelled (e.g. timeout) but the request is in
the process of being sent then a race can occur between the sending
and calling of the `onCompletion` request callback.

Changes
-------
Move circuit breakers out of memdpipeline client and onto memdclient.
This enables removing of the onCompletion callback as the logic
can be handled in the memdclient CancelRequest function. Also update
internalCancel to take an error so that we always provide errors to
the circuit breaker.

Change-Id: Id9d989607fe06c27ecdc59a7787b734d6937641d
Reviewed-on: http://review.couchbase.org/123299
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>