Change how managers/components register with config manager
Motivation
----------
Instead of agent controlling wiring managers and components up
to receive from the config manager we should instead give them
references to the config manager and allow them to register/
unregister themselves.

Changes
-------
Update the looper controllers to be able to directly push configs
to the config manager.
Update other managers and components to take a reference to
the config manager and add themselves as watchers.
Split out a waitUntil component to manage the waitUntil behaviour.

Change-Id: Ib77f243cb2676195163beb8cf1a2960179b2c5f1
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125368
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>