Get collections and durability support from ns config
Motivation
----------
We should get durability and collections support from the cluster
config rather than from the memdclient hello. This will allow
us to reset connections in the future if a cluster becomes
collections aware.

Changes
-------
Move ownership of collections and durability support to the
kv muxer state.
Add bucket capabilities to the route config and carry it over from the
cluster config.

Change-Id: I5d85590d61a5b64deba22d2f5997f07453aa07ff
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125370
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>