Move err map management to a manager
Motivation
----------
Handling of the error map is going to need to move so that retry
handling logic can also be moved to the kvmux. If we make this
a component then we can supply the kvmux with the component and
set it up listen to changes from memdclient bootstrapping.

Changes
-------
Move error map handling to a new error map manager.

Change-Id: I7f30dc40afc2d77bfaded4afe8ca5ebbecb1cf6b
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125377
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>