Move bootstrap to memdclient
Motivation
----------
Bootstrap should be owned by the memdclient as that's what
bootstrap is actually setting up.

Changes
-------
Move bootstrap to memdclient and create a memdclient dialer
component which is used by the kvmux to create and setup
new memdclient.
Remove properties from agent that don't need to be on it.

Change-Id: Ib1341de8cacee6eed3d2f275711384ce9dff5219
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/125382
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>